Add title to plot
Not sure if doing smearing correctly. Currently, smearing is done by taking the 1/root(num photons counted in chamber) and then calculating a random smear. So the smearing only occurs where there was either a noise photon or a true photon. So it doesn't work where there were 0 recorded events, even if those with 0 recorded events are right next to huge spikes of photons recorded. In those cases, the photon value is 0 for that chamber, not really like real-life detectors where the energy can spill over into adjacent chambers.