Start work on unfolding methods, stub for detector matrix
Not sure how to calculate the detector matrix correctly. Since it is the probability of having some energy show up in adjacent bins, it should be able to be generated from the detector itself. But the photon counts come after a variety of different sources smear and distort the data, so how to exactly quantify that? For now, skipping it and trying to implement the unfolding aglorithms assuming a well-formed detector matrix