Misc work on basic unfolding and detector matrix
Currently gives horrible error on the unfolding compared to the actual
distribution.

One problem is that the detector matrix does not always have an inverse that numpy can calculate, almost all the time the diagonal of the matrix includes zeros and I have to add a small value there to get numpy to do the inverse. These values then give huge errors. Not sure best way to deal with this yet.

Also still looking into how gradient descent works and how to do it correctly for 2D data.