Some progress on unfolding
Currently, the summed signal and the true energy distribution are somewhat close, like they should be. But the Y_vector distribution is way off for some reason, even though it is calculated in the same way as the summed signal. This also makes the unfolding wildly off, either super too large or too small, and always with almost no particles. Have no clue as to what is causing the problem at the moment.