Testing unfolding graphs for inconsistency
When using a perfect detector and the raw true energy, the graphs of the unfolding and true spectrum were widly different. Now they match perfectly, although still need to figure out how to get widths and other values for the graphs of it all without resorting to taking it from the y vector.