Start on eigenvalue cutoff
When using simulation that has no smear and no noise, only the variation from the gamma functions, the unfolding is still quite off, but is pretty close, except for the last two unfolded values, which shoot up to, in one case, from 10^3 to suddenly 10^19 and 10^20. Hoping eigenvalue cutoff will help with that problem.