Add changing number of bins in detector response matrix
Also try different ways of testing. Epsilon tests work, even for matricies with determinants extremely close to 0 (e.g. -1x10^-181) but doesn't seem to work for the ones created by the Detector class with a determinant of about 0.003. Not sure what the case is or why the issues with it. Plotting is still wonky because the bar plots don't exactly align to reality, only roughly, better to check the output arrays more carefully instead.