SVD now gives correct sum
Although still includes a hugely negative number of particles in the first bin
So not sure what to make of that, but the sum being correct is a good sign?

Not sure if its the correct way of dotting the matricies though, basically unsure about it all. And seems too exact, like how to correct for it and error bars?