Split up steps in gradient and Hess Matrix calculations
Gradient now gives actual numbers instead of -inf and NaNs, but the Hesse-Matrix is still not working. Hessian ends up always being a 1 in the top corner, 4's all the way down, and then the last few elements in the diagonal have some other values around 4 or 1, everything else being 0.