Update parts of operations to fix matrix mult errors
For example, calculating C_prime now should work, gets the same response as the llh one in funfolding. In addition, this implementation vs the funfolding llh implementation are very similar, but still prodce widly different results. Something is still wrong with this one, most likely in the gradient and almost certainly in the Hessian sections, the log likeliehood seems to do almost the same as the one in funfolding, although in a different manner. Still working on trying to deduce what the differences actually are.