Fix some errors with Hess and delta A calculation
Still doesn't run at the moment, the shapes are wrong for the iteration. new_true is a 2 element array, not the 20 element array that is expected, so need some work on that. But hopefully getting closer, most of the rest of the code seems to be consistent with the LLHTiknovForLoop in funfolding, although done in a different way. Hoepfully this iterative approach is correct.

Biggest issue with understanding the math is how the (next - current) should be interpreted. For now, I am assuming that is means that essentially the change in f is what is needed there, hence the change_in_a taking the spot of those in the software compared to the actual equation. Might need to be revised later.