Basic iterative working somewhat
Does iterate towards the lower -log likelihood, but is stuck in a loop once it reaches what it finds to be the lowest value. Not sure it is the correct lowest value, but is a resonable amount of particles in the new "true" distribution. Although now the "true" distribution isn't the actual true distribution anymore, so not sure how helpful that is.

Converges quickly though. Still need to test and probably refactor parts to make it work with non-square binning.