Try using scipy.minimize
Commodeering the unfolding parameter to swtich between my iterative approach and scipy's. Currently scipy.optomize.minimize is not working well, within the constraints, the fits is as bad as it possibly could be. All are either at the max or 0, and if there is no constraints, then it gives negatve numbers for some of the bins. So most likely log-likeliehood is wrong.