Update equations for iterative approach
Somewhat works better now, although the unfolded result is still worse than the signal, it is not a flat line. There is an issue with, with no regularization, the Hessian cannot be inverted, as it is a singular matrix. So currently regularization is needed for the unfolding to run, and tau has to be fine tuned to make a reasonable plot. Not really working, but a step closer in the right direction.