Tweak iterative approach
Error might stem with the fact that the fit is losing particles every time through, not sure where, but even after the supposed normalization, it drops. For example, from 5000.0 to 3998. to 3048, and then breaks. The lower particles lowers the probability that it is correct, probably causing the issues with it breaking after 1 or two loops