Work on fixing errors with non-square matricies
Not completed yet, does not work great yet. Impetus is also because the llh mcmc returns a result 3 time less likely as the one in funfolding, with the funfolding one seeming to be correct. llh functions themselves seem to be the same, not sure what the difference is there though....