Add basic reading in of FACT data
Issues with getting the 'runs' key with read_hf5py from the gamma MC data, as well as gamma_recuts.hdf5 only working with the Pandas read_hdf and not with the other one, and so there is some issue on what to use. currently loads the test data though.