Work on singular values plot
Trying to get the true to match up with the measured binning. Having the same true binning for all of them should help, but the values are way too small for the singular values. Might need to get the binning of the true distribution from the models themselves, which might be harder.