Work on Singular Value Comparison Plot
Almost doneish. Besides the condition number being incredibly large, the results seem mostly okay, with the tree binning outperforming the other binning, like expected. The downside of the current approach is that I am using a uniformly binned true distribution, which I am not sure how well that works to display the true singular values. Need to find out how to get the true binning from the ClassicBinning class to use in its outer product, and the same for the decision tree to use in its outer product for probably more accurate results.

In addition, still need to figure out what I need to use for y in the similar value plots so I can include them.